acs-package {acs}	R Documentation

Download, Manipulate, and Present American Community Survey and Decennial Data from the US Census

Description

Provides a general toolkit for downloading, managing, analyzing, and presenting data from the U.S. Census, including SF1 (Decennial "short-form"), SF3 (Decennial "long-form"), and the American Community Survey (ACS). Confidence intervals provided with ACS data are converted to standard errors to be bundled with estimates in complex acs objects. Package provides new methods to conduct standard operations on acs objects and present/plot data in statistically appropriate ways.

Details

Package: 	 acs

Type: 	 Package

Version: 	 2.1.3

Date: 	 2018-03-01

License: 	 GPL-3

Depends: 	 stringr, methods, XML

The package defines a new "acs" class (containing estimates, standard errors, geography, and metadata for tables from the U.S. Census American Community Survey), with methods to deal appropriately with common tasks, such as combining subgroups or geographies, mathematical operations on estimates, tests of significance, and computing (and plotting) confidence intervals.

Author(s)

Ezra Haber Glenn eglenn@mit.edu

References

"A Compass for Understanding and Using American Community Survey Data: What State and Local Governments Need to Know." Washington, DC: U.S. Census Bureau. 2009. http://www.census.gov/library/publications/2009/acs/state-and-local.html.

"acs.R: An R Package for Neighborhood-Level Data from the U.S. Census." Ezra Haber Glenn, Department of Urban Studies and Planning, Massachusetts Institute of Technology. Presented at the Computers in Urban Planning and Urban Management Conference, July 6, 2011. http://papers.ssrn.com/sol3/papers.cfm?abstract_id=2171390.

"Working with acs.R (June 2013)", Ezra Haber Glenn. http://papers.ssrn.com/sol3/papers.cfm?abstract_id=2552524

CityState webpage: http://eglenn.scripts.mit.edu/citystate/

User Group Mailing List: http://mailman.mit.edu/mailman/listinfo/acs-r

[Package acs version 2.1.3 Index]