acs.lookup-class {acs}	R Documentation

Class "acs.lookup"

Description

A new class to hold the results of calls to acs.lookup, typically for possible modification and then passing to calls to acs.fetch (using the "variable=" argument).

Objects from the Class

Objects can be created by calls of the form new("acs.lookup", ...), but more typically will be created as output from calls to acs.lookup.

Slots

endyear:

Object of class "numeric" indicating the year of the census dataset; e.g., for data from the 2005-2009 ACS, endyear=2009

span:

Object of class "numeric" indicating the span in years of the census dataset; e.g., for data from the 2005-2009 ACS, span=5. For decennial census datasets (SF1 and SF3), span = 0.

args:

Object of class "list" containing the search terms used in the call to acs.lookup, including some or all of: keyword, table.name, endyear, dataset, table.number, and case.sensitive.

results:

Object of class "data.frame" containing character strings in four columns: variable.code, table.number, table.name, and variable.name.

Methods

+

signature(e1 = "acs.lookup", e2 = "acs.lookup"): used for combining two acs.lookup objects into one

c

signature(x = "acs.lookup"): used for combining two acs.lookup objects into one

endyear

signature(object = "acs.lookup"): returns endyear from acs.lookup object

[

signature(object = "acs.lookup"): used for subsetting an acs.lookup object

results

signature(object = "acs.lookup"): returns results (as dataframe) from acs.lookup object

span

signature(object = "acs.lookup"): returns span from acs.lookup object

Author(s)

Ezra Haber Glenn eglenn@mit.edu

See Also

acs.lookup

Examples

showClass("acs.lookup")

[Package acs version 2.1.3 Index]