api.key.migrate {acs}	R Documentation

After updating the acs package, installs an archived API key from a previous installation.

Description

The acs.fetch function requires an "API key" to use when downloading data from the US Census API. Rather than pass this rather long string to the function each time, users can save the key as part of the package installation, using the api.key.install function. Once installed, an api key is saved on the system and available for use in future sessions. (To replace a key, simply call the function again with the new key.)

During the update process, this key may be lost or left in the wrong location. A call to api.key.migrate() can help restore an archived key, if found.

Usage

api.key.migrate()

Value

Migrates the key (if found) and exits silently; return an error if no archived key is found.

Author(s)

Ezra Haber Glenn eglenn@mit.edu

References

To request an API key, see http://www.census.gov/developers/

See Also

acs.fetch api.key.install

[Package acs version 2.1.3 Index]