cpi {acs}	R Documentation

Consumer Price Index data (1913-2015).

Description

Contains data on the Consumer Price Index for All Urban Consumers (CPI-U) for the years 1913-2015 from the U.S. Bureau of Labor Statistics. Used by the acs package for currency conversion functions. Scaled for base (100) to be 1982-84.

Usage

data(cpi)

Format

A named vector containing 103 observations, one for each year from 1913 through 2015.

Source

http://www.bls.gov/cpi/

See Also

currency.year

currency.convert

[Package acs version 2.1.3 Index]