endyear {acs}	R Documentation

Return or replace endyear value from the metadata of an acs object.

Description

endyear() will return the (integer) value of the latest year of the object (for example, for the 2005-2009 ACS survey, endyear = 2009.) When used for assignment, endyear<- will change the value of the endyear slot in an acs object, warning the user that this is an unusual thing to do.

Usage

endyear(object)

endyear(object)<-value

Arguments

object
an acs object

value
an integer to use as the new endyear

Details

Normal operations on acs objects should not involve altering the value of endyear (although users may wish to change the value of currency.year for comparisons with other objects). Sometimes endyear may be set incorrectly when data is imported, in which case endyear<- may be necessary.

Value

Returns (or replaces) an integer value from the endyear slot of an object.

Author(s)

Ezra Haber Glenn eglenn@mit.edu

See Also

currency.year, which is often what users will be intending to modify

acs-class

[Package acs version 2.1.3 Index]