fips.state {acs}	R Documentation

FIPS codes and geographic names for use in searching and creating geographies in the acs package.

Description

FIPS codes and geographic names for use in searching and creating geographies in the acs package. (Used internally.)

Usage

data(fips.state)

Format

Each table is a dataframe containing FIPS codes and names from the US Census.

Source

State: http://www.census.gov/geo/reference/ansi_statetables.html

County: http://www.census.gov/geo/www/codes/county/download.html

County Subdivision: http://www.census.gov/geo/www/codes/cousub/download.html

Place: http://www.census.gov/geo/www/codes/place/download.html

School: http://www.census.gov/geo/www/codes/sd/

American Indian Area: http://www.census.gov/geo/www/codes/aia/

[Package acs version 2.1.3 Index]