kansas09 {acs}	R Documentation

County-level data from the 2005-2009 American Community Survey for Kansas for use in examples of acs package.

Description

County-level data from the 2005-2009 American Community Survey for Kansas. Contains demographic data on sex, age, and citizenship. Used for examples in acs package. kansas09, and the corresponding one-year survey data in kansas07, provide acs objects to test and demonstrate various functions in the package.

Usage

data(kansas09)

Format

An acs-class object with 105 rows/geographies and 55 demographic variables, representing county-level ACS data for the state of Kansas. Also includes geographic and other metadata.

Source

U.S. Census American Community Survey, 2009; http://www.census.gov/

Examples

data(kansas09)
str(kansas09)
class(kansas09)

geography(kansas09)

# subsetting
kansas09[1:3,2:4]

# row-wise addition
kansas09[1,6]+kansas09[2,6]

# column-wise addition
kansas09[1:4,3]+kansas09[1:4,27]

[Package acs version 2.1.3 Index]