rbind.acs {acs}	R Documentation

Combine acs Objects by Rows

Description

Take a pair of acs objects and combine by rows.

Usage

          ## S3 method for class 'acs'
rbind(e1, e2, ...)
          

Arguments

e1, e2
two acs-class objects

...
provided for consistency with cbind S3 method

Details

When passed two acs-class objects, rbind will first check to confirm whether the objects contain compatible data: same endyear and span; same column names. If not, it will issue a warning, but will still proceed.

After this check, the function will return a new acs object that has resulted from combining the two arguments row-wise. The effect is essentially the same as rbind on the underlying estimate and standard.error matrices, with all the additional acs metadata tended to.

Value

Returns a single new acs object with all of the data contained in the two arguments.

Author(s)

Ezra Haber Glenn eglenn@mit.edu

[Package acs version 2.1.3 Index]